requirements.txt
requests==2.31.0
openai==1.3.9
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Flask==2.3.0
Flask-CORS==4.0.0
APScheduler==3.10.4
